refactor(metadata): one file_block_refs implementation over both SQL dialects
Every statement the manifest delta runs is identical in both dialects
apart from its placeholders, so the two copies collapse onto sql.Core
and Dialect gains Placeholder for the statements that are assembled at
runtime from a variable-length offset list and so have no constant to
bake.

Postgres loses its pgx.Batch and ANY() array form for the batched
IN-list and multi-row insert sqlite already used, which both dialects
accept. Below 180 changed chunks that is one statement where it used to
queue one per ref.

Fixes a drift the duplication was hiding: sqlite's manifest load never
selected start_offset, so every ChunkRef it returned carried zero. Only
FindByObjectID reads it, and its consumer is not built yet, so the wrong
offset was never reachable.
