+++ pkg/metadata/store/sql/executor.go
@@ -7,10 +7,12 @@
 // bodies riding on it can live in one place instead of two.
 //
 // What crosses this seam is the executor and error classification — the things
-// that genuinely differ at runtime. SQL *text* does not: ?N versus $N, NOW()
-// versus CURRENT_TIMESTAMP and the handful of rewritten queries stay as
-// per-dialect constants, because a Placeholder(n) method would move the same
-// two-line difference into an indirection and buy nothing. The consequence to be
+// that genuinely differ at runtime. SQL *text* mostly does not: ?N versus $N,
+// NOW() versus CURRENT_TIMESTAMP and the handful of rewritten queries stay as
+// per-dialect constants, because moving a two-line difference into an
+// indirection would buy nothing. The one text difference that does cross is
+// Dialect.Placeholder, for the statements assembled at runtime from a
+// variable-length list, which have no constant to bake. The consequence to be
 // honest about is that this merges logic, not SQL: what costs us maintenance is
 // hand-maintained control flow duplicated around the queries, not the query text.
 //
