+++ pkg/metadata/store/sqlite/dialect.go
@@ -3,6 +3,7 @@ package sqlite
 import (
 	"database/sql"
 	"errors"
+	"strconv"
 
 	"github.com/marmos91/dittofs/pkg/metadata/lock"
 	storesql "github.com/marmos91/dittofs/pkg/metadata/store/sql"
@@ -377,3 +378,6 @@ var blockRecordQueries = storesql.BlockRecordQueries{
 		WHERE block_id = ?2
 		RETURNING live_chunk_count`,
 }
+
+// Placeholder implements storesql.Dialect.
+func (dialect) Placeholder(i int) string { return "?" + strconv.Itoa(i) }
