+++ pkg/metadata/store/sqlite/transaction.go
@@ -283,7 +283,7 @@ func (tx *sqliteTransaction) putFile(ctx context.Context, file *metadata.File, w
 		// manifest, so putFileChunkRefs writes nothing and reports wrote=false.
 		// Freshly-inserted rows (!updated) have no prior refs, so every ref is
 		// a plain insert. The counter tracks manifests that truly changed.
-		wrote, scanned, err := putFileChunkRefs(ctx, tx.tx, file.ID, file.Blocks, updated, file.ManifestDirtyOffsets)
+		wrote, scanned, err := tx.PutFileChunkRefs(ctx, file.ID, file.Blocks, updated, file.ManifestDirtyOffsets)
 		tx.store.manifestRowsScanned.Add(int64(scanned))
 		if err != nil {
 			return mapDBError(err, "SetManifest", "blocks")
